Remove  most instances of `aria-hidden="true"`
The only real ‘presentational’ thing in these templates is the spacer
row, so I added this attribute on that <td> and removed it every else.
fixes #117

I also formatted the tabs/spaces/line-breaks which makes this commit
look way bigger than it really is. Shoulda done that in the separate
commit ¯\_(ツ)_/¯.