Add new component to Hybrid template: Background Image with Text, Icon, Button
This component is a more advanced version of Background Image with Text, whereby the colors of the background and background image are set up such that spacing issues will be revealed (for example if the background image does not fill the entire container). The only known issues with this component is that on Outlook, the height of the section is fixed, and thus must be manually adjusted based on the length of the content. Another known issue is that dark mode can cause the text to become harder to read (because it gets inverted but the background image stays the same).